1.Учел замечания по Matrix 2. Исправил ArrayListHome - добавил using

+++ CourseTasks/Vector/Vectors.cs
@@ -63,7 +63,7 @@ namespace Vectors
             }
         }
 
-        public double GetSize()
+        public int GetSize()
         {
             return vector.Length;
         }
